Initial commit
This is a working if incomplete scratchpad application; we can add
links, which are fetched (using `ureq`) and parsed with the Mozilla
Readability algorithm (using `dom_smoothie`), as well as adding freeform
notes, either to a link item or as a standalone item. Notes and links
may be tagged, and some tag normalization is applied.

In its current form, meowpad is essentially a write-only application; we
have not yet implemented `ls`, `show`, or `search` commands.