Upgrade/downgrade secondary links
This changeset add support for "casting" a link from secondary to
primary (if, for example, you wish to capture as searchable the
discussion on Lobsters or MetaFilter), as well as adding a related link
that already exists as a primary link. Removing a primary link does
not touch it anywhere it exists as a related link.