Update subcommand
This changeset adds the `update` subcommand for adjusting an existing
link. Five options are available: add-related-link; remove-related-link;
add-tags; remove-tags; and refresh. The last, `refresh`, will re-fetch
the link and update the title, description, and link content based on
the latest version at the URL.

Additionally, a display issue in the `show` subcommand's treatment of
related links was fixed.